Fix COALESCE type conflict and add autoBuyConfigs integration tests
$5 was used as both text (CorpSAG concatenation) and integer
(COALESCE with 0) in the corp division deficit query, causing
"COALESCE types text and integer cannot be matched". Cast $5::integer
explicitly in the COALESCE.

Added 12 integration tests covering all AutoBuyConfigs repository
methods: CRUD operations and all 4 GetStockpileDeficitsForConfig
code paths (character/corp × hangar/container).

+++ internal/repositories/autoBuyConfigs.go
@@ -284,7 +284,7 @@ func (r *AutoBuyConfigs) GetStockpileDeficitsForConfig(ctx context.Context, conf
 					AND sm.owner_id = $3
 					AND sm.location_id = $4
 					AND sm.container_id IS NULL
-					AND COALESCE(sm.division_number, 0) = COALESCE($5, 0)
+					AND COALESCE(sm.division_number, 0) = COALESCE($5::integer, 0)
 				GROUP BY sm.type_id, sm.desired_quantity, sm.price_source, sm.price_percentage
 			`
 		}
